4bar/requirements.txt
alabaster==0.7.12
alembic==1.0.0
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
colorama==0.4.0
DateTime==4.2
Django==2.1.2
docutils==0.14
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.7
image==1.5.24
imagesize==1.1.0
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
packaging==18.0
Pillow==5.2.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
requests==2.20.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.10
urllib3==1.24
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.5.0
